chore: add helper methods to read environment variables
